pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dba-assistant"
version = "0.1.0"
description = "Phase-oriented scaffold for the DBA Assistant project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents",
    "PyYAML>=6,<7",
    "python-docx>=1.1,<2",
    "redis>=5",
]

[project.scripts]
dba-assistant = "dba_assistant.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8,<9"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
